Data Acquisition Layout
Created a cmake for the STM32 folder

Laid out the initial steps we need for the data collection/acquisition from the mems mics.

Created guides for what needs to be done with CubeMX and the microphone setup.

Skeleton of the data acquisition is done, same with the organization of the repo. Need to populate the rest (including main)